+++ ItemStringParser/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.0.0
+
+- Add the `availableOnly`-Parameter
+- Change the signature
+- Add `ParseItemStringStrict()`
+- Droptables now also check if the items are available in the run
+
 # 1.0.1
 
 - Small readme.md fixes